Crear checktable para simplificar ModelTrait
